+++ udata/tasks.py
@@ -174,6 +174,7 @@ def init_app(app):
     import udata.core.badges.tasks  # noqa
     import udata.core.storages.tasks  # noqa
     import udata.features.notifications.tasks  # noqa
+    import udata.notifications.tchap  # noqa
     import udata.harvest.tasks  # noqa
     import udata.db.tasks  # noqa